feat: Move to lib and add tests
