+++ lib/index.test.js
@@ -1,0 +1,33 @@
+const lib = require('./index');
+const packageurl = require('packageurl-js');
+
+describe("replace version escape", () => {
+    test("replace @ in namespace", () => {
+        // https://www.npmjs.com/package/@angular/cli
+        const purl = "pkg:NPM/@angular/cli@4.17.21";
+
+        var new_purl = lib.replaceVersionEscape(purl);
+        expect(new_purl).toBe("pkg:NPM/%40angular/cli@4.17.21");
+        packageurl.PackageURL.fromString(new_purl);
+    });
+    test("if encoding ha already happened", () => {
+        const purl = "pkg:npm/es-abstract%401.16.0";
+
+        var new_purl = lib.replaceVersionEscape(purl);
+        expect(new_purl).toBe("pkg:npm/es-abstract@1.16.0");
+        packageurl.PackageURL.fromString(new_purl);
+
+        const purl2 = "pkg:npm/%40vue/cli-shared-utils%404.0.4";
+        var new_purl2 = lib.replaceVersionEscape(purl2);
+        expect(new_purl2).toBe("pkg:npm/%40vue/cli-shared-utils@4.0.4");
+        packageurl.PackageURL.fromString(new_purl2);
+
+    })
+    test("replace ^ in version", () => {
+        const purl = "pkg:NPM/@angular/cli@^4.17.21";
+
+        var new_purl = lib.replaceVersionEscape(purl);
+        expect(new_purl).toBe("pkg:NPM/%40angular/cli@%5E4.17.21");
+        packageurl.PackageURL.fromString(new_purl);
+    })
+})
