+++ package.json
@@ -31,8 +31,8 @@
     "@github/dependency-submission-toolkit": "2.0.4"
   },
   "devDependencies": {
-    "eslint": "^8.40.0",
     "@vercel/ncc": "^0.36.1",
-    "jest": "^29.4.2"
+    "eslint": "^8.40.0",
+    "jest": "^29.7.0"
   }
 }
